+++ controllers/userController.js
@@ -2,6 +2,9 @@ const userDAO = require('../dao/userDAO');
 
 const getAllUsers = async (req, res) => {
     try {
+        if (req.user.role != 1) {
+            return res.status(403).json({ message: 'You do not have permission' });
+        }
         const users = await userDAO.getAllUsers();
         res.status(200).json(users);
     } catch (error) {
@@ -11,6 +14,11 @@ const getAllUsers = async (req, res) => {
 
 const getUserById = async (req, res) => {
     try {
+
+        if (req.user.role != 1 && req.user.id != req.params.id) {
+            return res.status(403).json({ message: 'You do not have permission' });
+        }
+
         const user = await userDAO.getUserById(req.params.id);
         if (user) {
             res.status(200).json(user);
@@ -22,6 +30,8 @@ const getUserById = async (req, res) => {
     }
 };
 
+
+//TODO es
 const createUser = async (req, res) => {
     try {
         const { username, password, email, role, status } = req.body;
@@ -33,8 +43,15 @@ const createUser = async (req, res) => {
     }
 };
 
+
 const updateUser = async (req, res) => {
     try {
+
+        //TODO dodac cos podobnego dla uzytkownika
+        if (req.user.role != 1) {
+            return res.status(403).json({ message: 'You do not have permission' });
+        }
+
         const { id } = req.params;
         const { username, password, email, role, status } = req.body;
         const userDTO = new userDTO(id, username, password, email, role, status);
@@ -51,6 +68,10 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
     try {
+        if (req.user.role != 1) {
+            return res.status(403).json({ message: 'You do not have permission' });
+        }
+
         const deletedUser = await userDAO.deleteUser(req.params.id);
         if (deletedUser.changes > 0) {
             res.json({ message: 'User deleted successfully' });
